Last two abilities in Soul
NO ULTIMATE

But yeah Guarding Aura and Anima were both really easy thanks to the elbow grease put into their single target variations, with very few, if not any bugs to fix.

There is a weird bug with stacking shields, so I'll have to look into that wayyyyy later or when I have a programmer on the team.

anywho the spellbook is in a decent enough spot for a demo vid