Minor Fixes + Balancing
Changed the weights on the healing formula so the numbers are actually reasonable.

Fixed a weird targetting bug when you kill npcs (all I needed to do was change a target ref from a pure get to a validated get...)

Solving some player generation bugs and organized the code a little.