Replication
Got Projectiles to replicate but not show their damage numbers. Took a stab at player generation but that led to a dead end. A lot of rewriting on how player characters are generated will be needed. Until then, I'm going to return to Buffs, Debuffs, and AoE attacks and their network replication.